# openapi.yaml
openapi: 3.0.3
info:
  title: panorama-builder API
  description: |
    JSON API for compositing flat photos into a 360° panorama. Stores
    stations (camera setups), photos (with embedded pose), image
    measurements (reticle anchors on photos), and control points
    (cross-station landmarks with a user-seeded location estimate).

    Image measurements link to a control point; the CP is the latent
    real-world feature each observation describes.

    All resources use server-assigned 13-character base32 IDs.
    Authentication is out of scope — anyone with the URL has read+write
    access. Field-level validation lives in handler code; this spec
    documents shapes only.
  version: 1.0.0
servers:
  - url: /api
...
paths:
...
  /photos/{id}:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
    get:
      tags: [photos]
      summary: Get photo metadata
      operationId: getPhoto
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [photos]
      summary: Partial-update photo pose / opacity / aspect / locks
      description: |
        Partial update: only the keys present in the request body are
        written. Omitting a key preserves its current value.
      operationId: updatePhoto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhotoUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [photos]
      summary: Delete a photo (and its blob)
      operationId: deletePhoto
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /photos/{id}/blob:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
    put:
      tags: [photos]
      summary: Upload the photo's image blob
      description: |
        Stores the uploaded bytes content-addressed at `blobs/<sha256>` and
        writes the resulting path into the photo's `blob_path`. Re-uploading
        different bytes produces a new hash; the previous bytes remain on
        disk and are still referenced by any prior commit's view of the row.
      operationId: uploadPhotoBlob
      requestBody:
        required: true
        content:
          'image/*':
            schema:
              type: string
              format: binary
      responses:
        '204': { description: Stored }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: Payload too large (above MAX_BLOB_BYTES)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [photos]
      summary: Download the photo's image blob
      description: |
        Resolves the photo's `blob_path` (overlay-aware when `X-Session-Id`
        is supplied) and streams the bytes. Callers that already have the
        content hash should prefer `GET /blobs/{hash}`, which serves the
        bytes directly with immutable cache headers and works without a
        session lookup.
      operationId: getPhotoBlob
      responses:
        '200':
          description: OK
          content:
            'image/*':
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFound' }
  /blobs/{hash}:
    parameters:
      - name: hash
        in: path
        required: true
        schema: { type: string, pattern: '^[0-9a-f]{64}$' }
    get:
      tags: [photos]
      summary: Download an image blob by its content hash
      description: |
        Content-addressed blob fetch. Bytes are immutable; the response is
        marked `Cache-Control: public, max-age=31536000, immutable`.
        Content-Type is sniffed from the first 512 bytes.
      operationId: getBlob
      responses:
        '200':
          description: OK
          content:
            'image/*':
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
...
    PhotoId:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Id:
      type: string
      description: 13-character base32 server-assigned id
      pattern: '^[A-Z2-7]{13}$'
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Photo:
      type: object
      required:
        - id
        - station_id
        - blob_path
        - mime_type
        - size_bytes
        - aspect
        - photo_az
        - photo_tilt
        - photo_roll
        - size_rad
        - opacity
        - lock_photo_az
        - lock_photo_tilt
        - lock_photo_roll
        - lock_size_rad
        - dist_k1
        - dist_k2
        - lock_dist_k1
        - lock_dist_k2
        - created_at
        - updated_at
      properties:
        id: { $ref: '#/components/schemas/Id' }
        station_id: { $ref: '#/components/schemas/Id' }
        blob_path:
          type: string
          nullable: true
          description: |
            Relative path under STORAGE_DIR where the photo's bytes live.
            New uploads land at `blobs/<sha256>` (content-addressed,
            immutable). Legacy rows may still carry `photos/<id>` until the
            startup migration rewrites them.
        mime_type: { type: string, nullable: true }
        size_bytes: { type: integer, format: int64, nullable: true }
        aspect: { type: number, format: double, description: width / height of the photo in pixels }
        photo_az: { type: number, format: double, description: viewer-azimuth radians (CCW from -Z) }
        photo_tilt: { type: number, format: double, description: viewer-altitude radians (above horizon) }
        photo_roll: { type: number, format: double, description: in-plane roll radians }
        size_rad: { type: number, format: double, description: horizontal angular subtense radians }
        opacity: { type: number, format: double, minimum: 0, maximum: 1 }
        lock_photo_az: { type: boolean }
        lock_photo_tilt: { type: boolean }
        lock_photo_roll: { type: boolean }
        lock_size_rad: { type: boolean }
        dist_k1: { type: number, format: double, description: Brown-Conrady radial distortion coefficient k1; 0 = no distortion }
        dist_k2: { type: number, format: double, description: Brown-Conrady radial distortion coefficient k2; 0 = no distortion }
        lock_dist_k1: { type: boolean, description: locked at zero by default; unlock to let the solver fit it }
        lock_dist_k2: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PhotoUpdate:
      type: object
      description: |
        Partial-update body for PUT /photos/{id}. Every field is
        optional; only keys present in the body are written.
      properties:
        aspect: { type: number, format: double }
        photo_az: { type: number, format: double }
        photo_tilt: { type: number, format: double }
        photo_roll: { type: number, format: double }
        size_rad: { type: number, format: double }
        opacity: { type: number, format: double, minimum: 0, maximum: 1 }
        lock_photo_az: { type: boolean }
        lock_photo_tilt: { type: boolean }
        lock_photo_roll: { type: boolean }
        lock_size_rad: { type: boolean }
        dist_k1: { type: number, format: double }
        dist_k2: { type: number, format: double }
        lock_dist_k1: { type: boolean }
        lock_dist_k2: { type: boolean }
...